requirements.txt
Flask
Flask_RESTful
requests
websocket-client
